Using castStyle
